Move tests from test_cuda_array_interface
